BIG UPDATE: - NBS-Predict is now in v1.0.0-alpha2. - Fix several bugs in the GUI. - Fix bug in correlation metric. - Change default parameters for searching algorithms. Now they have more liberal parameters (e.g., reduced number of iterations). - Add function to control confounds. - Add a menu in GUI to choose a scaling method. - SVM Regression is now built-in function instead of LIBSVM. - Add Linear Regression, Logistic Regression and Linear Discriminant Analysis - Add MinMaxScaler, StandardScaler and MaxAbsScaler - LIBSVM is now removed. Instead, builtin SVM functions are now available to use. They give the same prediction performance as LIBSVM, but are significantly faster. - Add preprocessing step in shrinkMat function, that features where information is less than 10% are removed. In this way, very sparse features are now removed. It might be increased to 20% in future updates.

+++ src/helper/modelFitScore.m
@@ -33,5 +33,5 @@ for i = 1:nMetrics
     varargout{i} = Mdl.score(trainTestData.y_test,...
         y_pred,metrics{i}); % score
 end
-varargout{i+1} = {uint8(trainTestData.y_test),uint8(y_pred)};
+varargout{i+1} = {single(trainTestData.y_test),single(y_pred)};
 end
